+++ cypress/integration/address_book.spec.js
@@ -20,6 +20,7 @@ describe('Address book', () => {
   it('should add and remove Address Book entries', () => {
     cy.visit(`/${RINKEBY_TEST_SAFE}/address-book`)
 
+    // Add a new entry manually
     cy.findByText('Create entry', { timeout: 6000 }).click()
     cy.findByTestId('create-entry-input-name').type(NAME)
     cy.findByTestId('create-entry-input-address').type(ENS_NAME)
@@ -28,28 +29,34 @@ describe('Address book', () => {
     cy.findByText(NAME).should('exist')
     cy.findByText(ADDRESS).should('exist')
 
+    // Edit the entry
     cy.get('[title="Edit entry"]').click({ force: true }) //This is because the button is hidden
     cy.findByTestId('create-entry-input-name').clear().type(EDITED_NAME)
     cy.findByTestId('save-new-entry-btn-id').click()
     cy.findByText(NAME).should('not.exist')
     cy.findByText(EDITED_NAME).should('exist')
 
+    // Delete the entry
     cy.get('[title="Delete entry"]').click({ force: true }) //This is because the button is hidden
     cy.findByText('Delete').should('exist').click()
     cy.findByText(EDITED_NAME).should('not.exist')
+    cy.get('.MuiSnackbarContent-action').click({ multiple: true }) // close the snackbar
 
-    cy.wait(4000) // Waiting for notifications to dissapear before clicking "Import"
+    // Import CSV
     cy.get('[data-track="address-book: Import"]').click()
     cy.get('[type="file"]').attachFile('../fixtures/address_book_test.csv')
     cy.get('.modal-footer').findByText('Import').click()
     cy.findByText(RINKEBY_CSV_ENTRY.name).should('exist')
     cy.findByText(RINKEBY_CSV_ENTRY.address).should('exist')
+
+    // Go to a Safe on Gnosis Chain
     cy.get('nav').findByText('Rinkeby').click()
     cy.findByText('Gnosis Chain').click()
     cy.visit(`/${GNO_TEST_SAFE}/address-book`)
     cy.findByText(GNO_CSV_ENTRY.name).should('exist')
     cy.findByText(GNO_CSV_ENTRY.address).should('exist')
 
+    // Download the export file
     const date = new Date()
     const day = `00${date.getUTCDate()}`.slice(-2)
     const month = `00${date.getUTCMonth() + 1}`.slice(-2)
